services: service metric log structure is more granular

+++ src/lib/work_distributor/driver/redis/event.go
@@ -6,7 +6,7 @@ const (
 )
 
 type RedisCommandExecutedEvent struct {
-	Count int8
+	Count int
 }
 
 type RedisLockReleasedEvent struct {
